# docs/swagger.yaml
basePath: /
definitions:
  cmd.IDResponse:
    properties:
      id:
        example: a3509860-d66f-4be4-8984-0b7a15b8f10c
        type: string
    type: object
  cmd.TotalResponse:
    properties:
      total:
        example: 100500
        type: integer
    type: object
  cmd.subscriptionCreateBody:
    properties:
      end_date:
        example: ""
        type: string
      price:
        example: 400
        type: integer
      service_name:
        example: Yandex Plus
        type: string
      start_date:
        example: "2025-07-01"
        type: string
      user_id:
        example: 60601fee-2bf1-4721-ae6f-7636e79a0cba
        type: string
    type: object
  cmd.subscriptionUpdateBody:
    properties:
      end_date:
        example: ""
        type: string
      id:
        example: '-'
        type: string
      price:
        example: 400
        type: integer
      service_name:
        example: Yandex Plus
        type: string
      start_date:
        example: "2025-07-01"
        type: string
      user_id:
        example: 60601fee-2bf1-4721-ae6f-7636e79a0cba
        type: string
    type: object
  github_com_edzh1_rest-effective-mobile_internal_models.Subscription:
    properties:
      end_date:
        type: string
      id:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
host: localhost:3000
info:
  contact: {}
  description: Тестовое задание по управлению подписками.
  title: rest-effective-mobile/
  version: "1.0"
paths:
  /subscriptions:
    get:
      consumes:
      - application/json
      description: Get list of subscriptions with optional filters
      parameters:
      - description: User ID filter
        format: uuid
        in: query
        name: user_id
        type: string
      - description: Service name filter
        in: query
        name: service_name
        type: string
      - description: Start date filter (YYYY-MM-DD)
        format: date
        in: query
        name: start_date
        type: string
      - description: End date filter (YYYY-MM-DD)
        format: date
        in: query
        name: end_date
        type: string
      - description: Page number
        in: query
        maximum: 100
        minimum: 1
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_edzh1_rest-effective-mobile_internal_models.Subscription'
            type: array
        "400":
          description: Invalid parameter format
          schema:
            type: string
      summary: List subscriptions with filters
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Create a new subscription record
      parameters:
      - description: Subscription data
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/cmd.subscriptionCreateBody'
      produces:
      - application/json
      responses:
        "200":
          description: '{"id": "a3509860-d66f-4be4-8984-0b7a15b8f10c"}'
          schema:
            $ref: '#/definitions/cmd.IDResponse'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Create new subscription
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a subscription by ID
      parameters:
      - description: Subscription ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Invalid UUID format
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Delete subscription
      tags:
      - subscriptions
    get:
      consumes:
      - application/json
      description: Get a single subscription by its ID
      parameters:
      - description: Subscription ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_edzh1_rest-effective-mobile_internal_models.Subscription'
        "400":
          description: Invalid UUID format
          schema:
            type: string
        "404":
          description: 404 page not found
          schema:
            type: string
      summary: Get subscription by ID
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: Update an existing subscription
      parameters:
      - description: Subscription ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Updated subscription data
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/cmd.subscriptionUpdateBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd.IDResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Update subscription
      tags:
      - subscriptions
  /subscriptions/total:
    get:
      consumes:
      - application/json
      description: Calculate total cost of subscriptions for a period with filters
      parameters:
      - description: User ID filter
        format: uuid
        in: query
        name: user_id
        type: string
      - description: Service name filter
        in: query
        name: service_name
        type: string
      - description: Period start (YYYY-MM-DD)
        format: date
        in: query
        name: start_date
        type: string
      - description: Period end (YYYY-MM-DD)
        format: date
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd.TotalResponse'
        "400":
          description: Invalid parameter format
          schema:
            type: string
      summary: Calculate total subscription cost
      tags:
      - subscriptions
swagger: "2.0"